quality: clean compile warnings + emit compile_commands.json
Two real warnings emitted by the toolchain:

1. tools/render_presets/main.cpp:157 — implicit signed→unsigned in
   wav.createWriterFor() call. Now passes int→double for sample rate
   and int→unsigned int for channel count explicitly.

2. tools/render_presets/main.cpp:164 — implicit int / unsigned long
   division. `interleaved.size() / channels` now casts channels to
   std::size_t.

Plus: CMAKE_EXPORT_COMPILE_COMMANDS is now ON at the top level, with a
`compile_commands.json` symlink at the repo root pointing at
`build/compile_commands.json`. Stops clangd / VS Code IntelliSense
from emitting "header not found" diagnostics on every file because
it didn't know the CMake include paths. The symlink path is in
.gitignore (added Phase 0).

After this: `cmake --build build --parallel` emits no compile
warnings (the only remaining noise is JUCE-internal ranlib reports
on translation units with no symbols, which we can't fix from our
side).

+++ tools/render_presets/main.cpp
@@ -154,14 +154,19 @@ void writeWav(const juce::File& dest, const std::vector<float>& interleaved, int
         std::fprintf(stderr, "WAV open failed: %s\n", dest.getFullPathName().toRawUTF8());
         return;
     }
-    auto* writer = wav.createWriterFor(stream.get(), sampleRate, channels, 32, {}, 0);
+    auto* writer = wav.createWriterFor(stream.get(),
+                                       static_cast<double>(sampleRate),
+                                       static_cast<unsigned int>(channels),
+                                       32,
+                                       {},
+                                       0);
     if (writer == nullptr)
     {
         std::fprintf(stderr, "WAV writer create failed\n");
         return;
     }
     stream.release(); // writer takes ownership
-    const int numFrames = static_cast<int>(interleaved.size() / channels);
+    const int numFrames = static_cast<int>(interleaved.size() / static_cast<std::size_t>(channels));
     juce::AudioBuffer<float> buffer(channels, numFrames);
     for (int ch = 0; ch < channels; ++ch)
     {
